rename settings -> application_settings
