+++ services/web/server/src/simcore_service_webserver/application_settings.py
